Fix missing dependencies; Dockerfile cleanup
When the stock image is run as a container, it fails noting a missing
dependency:

ansible.plugins.filters.ipaddr

To fix this, instead of installing python, I used apk to add 'ansible'
which contains all of the dependencies needed.

Then, I cleaned up the remaining lint warnings in the Dockerfile and
made the runner run as nobody instead of root.

Finally, I added a Bash script to add a function to the current
environment that runs the linter for us without having to use the
Docker-specific syntax; rather:

j2lint /path/to/filename /path/to/other/filename

The function iterates through each of the files passed (so, globbing and
such still works) mounts the directory containing the file to be linted,
runs the linter on the file relative to /check/ , then rewrites the
output so that the original directory names are restored.  Therefore,
if I ran `j2lint /path/to/filename` , the results will show
'/path/to/filename' instead of '/check/filename' which ought to help
with readability.